refactor: removed 'removeRefPrefix()' and component.ts, re-factored removeXtensions() to be a pure function
