+++ packages/connector/src/index.ts
@@ -1,5 +1,4 @@
 export * from './types.js';
-export * from './components.js';
 export * from './parseParams.js';
 export * from './createSecurityProcessors.js';
 export * from './routeSchema.js';
